+++ chem_data/data_set.py
@@ -1,5 +1,5 @@
 import pandas as pd
 def load_data_frame():
-    dataFrame = pd.read_csv('../data/molecule_data.csv')
+    dataFrame = pd.read_csv('data/molecule_data.csv')
     dataFrame['dH'] = dataFrame['dH'].astype(float)
     return dataFrame
