+++ test/test_balancing_equations.py
@@ -1,4 +1,4 @@
-from repository.balancing_equations import *
+from chem_calculations.balancing_equations import *
 #Test method for entering formulas
 #Checks to see input is valid.
 def test_entering_formulas():
