+++ test/test_chem_ui.py
@@ -1,0 +1,6 @@
+from chem_textual_ui import *
+from chem_textual_ui.chem_ui import ChemApp
+def start_chem_app():
+    app = ChemApp()
+    app.run()
+start_chem_app()
