--- ui/test_chem_ui.py
@@ -1,7 +1,0 @@
-from chem_ui import *
-def start_chem_app():
-    app = ChemApp()
-    app.run()
-
-
-start_chem_app()
